changed foler structure for intent and unit tests
