+++ app/src/androidTest/java/ca/cmput301t05/placeholder/profile/intentTestProfilePage.java
@@ -26,30 +26,9 @@ import ca.cmput301t05.placeholder.MainActivity;
 import ca.cmput301t05.placeholder.R;
 import ca.cmput301t05.placeholder.events.Event;
 
-public class ProfileTest {
 
-    public Profile mockProfile(){
-        return new Profile("Bob", UUID.randomUUID());
-    }
-
-    public Event mockEvent(){
-        return new Event("Soccer Game", "Bring water bottle", 40);
-    }
-    @Test
-    public void testJoinEvent() {
-        // Tests joining and leaving event functionality in profile
-        Profile userProfile = mockProfile();
-        Event event  = mockEvent();
-        userProfile.joinEvent(event);
-        assertEquals(userProfile.getJoinedEvents().size(), 1); // Joined one event should should return one event
-        userProfile.leaveEvent(event);
-        assertEquals(userProfile.getJoinedEvents().size(), 0);
-    }
-
-
-    @Nested
     @RunWith(MockitoJUnitRunner.class)
-    public static class intentTestProfilePage {
+    public class intentTestProfilePage {
         // Intent test. Different from unit test above.
 
         @Rule
@@ -64,7 +43,7 @@ public class ProfileTest {
             onView(withId(R.id.intro_name_edit)).perform(ViewActions.typeText("Dave"), closeSoftKeyboard()); // Entering name
             // Will go to the main activity homepage
             onView(withId(R.id.intro_submit_button)).perform(click());
-            SystemClock.sleep(5000);
+            SystemClock.sleep(2000);
             onView(withId(R.id.btnProfile)).perform(click()); // Click on profile button
             onView(ViewMatchers.withId(R.id.edit_name))
                     // Check that the text matches the expected name "Dave"
@@ -75,4 +54,3 @@ public class ProfileTest {
 
     }
 
-}
